requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
zhipuai==2.0.1
sentence-transformers==2.2.2
qdrant-client==1.9.0
torch==2.1.0
python-dotenv==1.0.0
requests==2.31.0
